+++ examples/butler/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"strings"
 
 	"go.bbkane.com/warg"
 	"go.bbkane.com/warg/value/scalar"
@@ -25,6 +26,8 @@ func app() *warg.App {
 					warg.EnvVars("BUTLER_PRESENT_NAME", "USER"),
 					warg.Required(),
 				),
+				warg.AllowForwardedArgs(),
+				warg.CmdFooter("Treat butler well, and he will serve you faithfully."),
 			),
 		),
 	)
@@ -35,6 +38,9 @@ func present(ctx warg.CmdContext) error {
 	// this is a required flag, so we know it exists
 	name := ctx.Flags["--name"].(string)
 	fmt.Fprintf(ctx.Stdout, "May I present to you %s.\n", name)
+	if len(ctx.ForwardedArgs) > 0 {
+		fmt.Fprintf(ctx.Stdout, "And: %s\n", strings.Join(ctx.ForwardedArgs, " "))
+	}
 	return nil
 }